+++ tests/TestYtDlpArgsBuilder.h
@@ -18,6 +18,7 @@ private slots:
     void testUnsupportedOpusThumbnailEmbedding();
     void testAudioPlaylistFolderJpg();
     void testAudioPlaylistArtistMetadataFallback();
+    void testAudioPlaylistIndexBecomesTrackMetadata();
     void testOrphanedTemporaryDirectorySweep();
     void testTemporaryDirectoryOwnershipGuard();
 };